% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\title{Subset __X coverage using Nanopore Reads}
\usage{
coverage(reads, genome_size_mb)
}
\arguments{
\item{reads}{A dataframe or list of dataframes containing the column names: read_id and sequence_length_template.}

\item{genome_size_mb}{Estimated genome size in megabases.}
}
\value{
A vector of subset read_ids/indexes whose length is equal to the desired coverage of a given genome.
}
\description{
Subset __X coverage using Nanopore Reads
}
\examples{
2x.subset <- nanopore_subset(reads, 2, 1500, TRUE, 10)
}
\author{
Auden Block
}
