% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorado_table.R
\name{dorado_table}
\alias{dorado_table}
\title{Dataframe containing a summary table about a Dorado run.}
\usage{
dorado_table(simplex, duplex = NA)
}
\arguments{
\item{simplex}{A simplex read dataframe.}

\item{duplex}{A duplex read dataframe from the same summary as the simplex dataframe.}
}
\value{
A subset of the input duplex dataframe in which a read is used as both a template and a complement. Useful for finding false positive duplex reads
}
\description{
Dataframe containing a summary table about a Dorado run.
}
\examples{
library2_duplex_table <- dorado_table(simplex.df, duplex.df)
all.libraries.duplex.table <- as.data.frame(lapply(simplex.list, dorado_table))
}
\author{
Auden Block
}
