% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplex_duplicates.R
\name{duplex.duplicates}
\alias{duplex.duplicates}
\title{Find the duplex reads in which a simplex read was used as a template for one duplex read, and a complement for a different duplex read.}
\usage{
duplex.duplicates(duplex)
}
\arguments{
\item{duplex.df}{A dataframe of duplex reads}
}
\value{
A subset of the input duplex dataframe in which a read is used as both a template and a complement.
}
\description{
Find the duplex reads in which a simplex read was used as a template for one duplex read, and a complement for a different duplex read.
}
\examples{
duplex.duplicates <- duplex.duplicates(duplex.df);
summary.list$duplex.duplicates <- duplex.duplicates(summart.list$duplex);
}
\author{
Auden Block
}
