% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplex_parents.R
\name{duplex.parents}
\alias{duplex.parents}
\title{Mutate the duplex dataframe to include information about the simplex parents from the simplex dataframe.}
\usage{
duplex.parents(duplex, simplex)
}
\arguments{
\item{duplex.df}{A dataframe of duplex reads}

\item{simplex.df}{A dataframe of simplex reads from which the duplex reads were generated}
}
\value{
The duplex dataframe with each duplex read now having the length and quality of its simplex parents.
}
\description{
Mutate the duplex dataframe to include information about the simplex parents from the simplex dataframe.
}
\examples{
duplex.with.parents.df <- duplex.parents.df(duplex.df, simplex.df);
summary.list$duplex <- duplex.parents.df(summart.list$duplex, summary.list$duplex);
}
\author{
Auden Block
}
