% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplex_percentage.R
\name{duplex.rate}
\alias{duplex.rate}
\title{Find either the total duplex READ percentage or duplex percentage over specific time intervals.}
\usage{
duplex.rate(duplex, simplex, interval = 0, type = "nucleotide")
}
\arguments{
\item{interval}{(Default = 0, the entire experimental run) Interval divisions in MINUTES to report the rate as a decimal value. For instance, 60 will return the rate of duplex reads generated every 60 minutes of run time.}

\item{type}{(default = "nucleotide") The type of rate to be reported: nucleotide or read.}

\item{duplex.df}{A dataframe of duplex reads}

\item{simplex.df}{A dataframe of simplex reads}
}
\value{
A dataframe of (,1:3) the interval evaluated reported in seconds, minutes, and hours, and (,4) the duplex rate as a decimal during that time interval.
}
\description{
Find either the total duplex READ percentage or duplex percentage over specific time intervals.
}
\examples{
simplex.parents <- simplex.parents(duplex.df, simplex.df);
summary.list$simplex.parents <- simplex.parents(summary.list$duplex.df, summary.list$simplex.df);
}
\author{
Auden Block

Dorado will automatically report the duplex rate over the entire run time as a nucleotide rate into the standard output (std.out). However, this can be difficult to utilize within R or if the std.out files are lost. In addition to a nucleotide rate duplex.rate can output the duplex read rate to evaluate the number of duplex reads generated.
}
