% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplex_parents.R
\name{duplex_parents}
\alias{duplex_parents}
\title{Mutate the duplex dataframe to include information about the simplex parents from the simplex dataframe.}
\usage{
duplex_parents(duplex, simplex, mutate = TRUE)
}
\arguments{
\item{duplex}{A dataframe of duplex reads}

\item{simplex}{A dataframe of simplex reads from which the duplex reads were generated}

\item{mutate}{(Default = True) Mutate an existing dataframe with the information, or return a new dataframe.}
}
\value{
Either the duplex dataframe with each duplex read now having the length and quality of its simplex parents (mutate = true) or a new dataframe of only the duplex parents.
}
\description{
Mutate the duplex dataframe to include information about the simplex parents from the simplex dataframe.
}
\examples{
duplex.with.parents.df <- duplex_parents(duplex.df, simplex.df);
summary.list$duplex <- duplex_parents(summary.list$duplex, summary.list$duplex);
duplex.parents.only.df <- duplex_parents(duplex.df, simplex.df, mutate = FALSE)
}
\author{
Auden Block
}
