% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplex_rate.R
\name{duplex_rate}
\alias{duplex_rate}
\title{Find either the total duplex percentage over a run.}
\usage{
duplex_rate(duplex, simplex, type = "nucleotide")
}
\arguments{
\item{type}{(default = "nucleotide") The type of rate to be reported: nucleotide or read.}

\item{duplex.df}{A dataframe of duplex reads}

\item{simplex.df}{A dataframe of simplex reads}
}
\value{
A dataframe of the run time evaluated in seconds, minutes, and hours, and the duplex rate as a decimal value.
}
\description{
Find either the total duplex percentage over a run.
}
\examples{
duplex_rate.df <- duplex_rate(duplex.df, simplex.df)
duplex_rate.df <- duplex_rate(duplex.df, simplex.df, type = "read")
}
\author{
Auden Block

Dorado will automatically report the duplex rate over the entire run time as a nucleotide rate at the end of basecaling. However, this can be difficult to utilize within R or recalculate if a standard output file is lost. We also included the ability to evaluate the number of duplex reads generated over the experiment.
}
