% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_function.R
\name{l_function}
\alias{l_function}
\title{Contiguity Quality.}
\usage{
l_function(read.length, percentage = 50)
}
\arguments{
\item{read.length}{A vector of the read length of a given dataset.}

\item{percentage}{(default = 50) The percentage at which to evaluate.}
}
\value{
The n.percentage of the given input dataframe.
}
\description{
Typically represented as the l.50, the function defines the count of contigs whoses sum length represents half the genome.
For more information, please see: https://ucdavis-bioinformatics-training.github.io/2020-Genome_Assembly_Workshop/kmers/QAQC
}
\examples{
n.function(summary.list$simplex$sequence_length_template);
n.function(sequence_length, 90);
n.function(c(5,3,55,24,6,3,1,5,7,9,75,3,2,12,6,45,3,565,57))
}
\author{
Auden Block
}
\concept{l.50}
