% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_function.R
\name{n_function}
\alias{n_function}
\title{Contiguity Quality.}
\usage{
n_function(read.lengths, percentage = 50)
}
\arguments{
\item{read.lengths}{A vector of the read lengths.}

\item{percentage}{(default = 50) The percentage at which to evaluate.}
}
\value{
The n.percentage of the given input dataframe.
}
\description{
Typically represented as the n.50, the function defines the contiguity quality of a given set of contigs. 
The n.function returns the sequence length of the shortest contig at the given percentage of the genome. For example, an n.50 represents the shortest read length to represent 50% of the genome. 
For more information, please see: https://ucdavis-bioinformatics-training.github.io/2020-Genome_Assembly_Workshop/kmers/QAQC
}
\examples{
n.function(summary.list$simplex$sequence_length_template);
n.function(sequence_length, 90);
n.function(c(5,3,55,24,6,3,1,5,7,9,75,3,2,12,6,45,3,565,57))
}
\author{
Auden Block
}
\concept{n.50}
