% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanopore_filter.R
\name{nanopore_filter}
\alias{nanopore_filter}
\title{Nanopore_filter}
\usage{
nanopore_filter(reads, min_quality = 1, min_length = 200, IDs = FALSE)
}
\arguments{
\item{reads}{A dataframe containing the column names: sequence_length_template, mean_qscore_template, and (if IDs=TRUE) read_ids.}

\item{min_quality}{(default = 1) Minimum acceptable quality score}

\item{min_length}{(default = 200) Minimum acceptable read length}

\item{IDs}{(default = false) By default, dataframe will be returned with the reads that pass filtering. If TRUE, a vector of read_IDs that passed quality scoring will be returned}
}
\value{
A dataframe of reads that pass quality and length filtering, or a character vector of readIDs that pass quality filtering.
}
\description{
Nanopore_filter
}
\examples{
library2_duplex_table <- dorado_table(simplex.df, duplex.df)
all.libraries.duplex.table <- as.data.frame(lapply(simplex.list, dorado_table))
}
\author{
Auden Block
}
