% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qscore_vs_readlength.R
\name{qscore_vs_readlength_graph}
\alias{qscore_vs_readlength_graph}
\title{Dataframe containing a summary table about a Dorado run.}
\usage{
qscore_vs_readlength_graph(
  reads,
  qscore = 10,
  histogram = FALSE,
  colors = c("#e76f51", "#2a9d8f"),
  theme = ggplot2::theme_classic()
)
}
\arguments{
\item{reads}{A read dataframe containing the columns "sequence_length_template" and "mean_qscore_template".}

\item{qscore}{(Default = 10) What value is an acceptable q-score.}

\item{histogram}{(Default = FALSE) If true, ggExtra will be used to create a histogram of the x- & y-axis.}

\item{colors}{If ggplot is active in the namespace, the colors used to differentiate between pass and fail. If using default R, the colors will unfortunately be random.}

\item{theme}{(Default = theme_classic()) When using ggplot2, what theme should the plot be formatted in.}
}
\value{
A subset of the input duplex dataframe in which a read is used as both a template and a complement. Useful for finding false positive duplex reads
}
\description{
Dataframe containing a summary table about a Dorado run.
}
\examples{
library2_duplex_table <- dorado_table(reads.df, duplex.df)
all.libraries.duplex.table <- as.data.frame(lapply(reads.list, dorado_table))
}
\author{
Auden Block
}
