% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fastq.R
\name{read_fastq}
\alias{read_fastq}
\title{Import a fastq file into R}
\usage{
read_fastq(fastq)
}
\arguments{
\item{fastq}{The path to the fastq file to be imported.}
}
\value{
A dataframe containing the raw header line (ReadID, duplex value, etc), sequence, base quality scores, and a column containing the parsed ReadIDs.
}
\description{
Import a fastq file into R
}
\examples{
fastq <- read_fastq("<path_to_file>")
}
\author{
Auden Block

This function is for those wishing to use R to manipulate and evaluate fastq files either through DuplexVisualized or other bioinformatics programs.
}
