% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_duplex_rate.R
\name{rolling_duplex_rate}
\alias{rolling_duplex_rate}
\title{Find either the rolling duplex percentage over specific time intervals.}
\usage{
rolling_duplex_rate(duplex, simplex, interval = 60, type = "nucleotide")
}
\arguments{
\item{interval}{(Default = 60) Interval in MINUTES to generate a rolling average. For instance, 60 will return the rate of duplex reads generated every 60 minutes of run time.}

\item{type}{(Default = "nucleotide") The type of rate to be reported: nucleotide or read.}

\item{duplex.df}{A dataframe of duplex reads}

\item{simplex.df}{A dataframe of simplex reads}
}
\value{
A dataframe of (,1:3) the interval evaluated reported in seconds, minutes, and hours, and (,4) the duplex rate as a decimal during that time interval.
}
\description{
Find either the rolling duplex percentage over specific time intervals.
}
\examples{
rolling_duplex_rate.df <- rolling_duplex_rate(duplex.df, simplex.df, interval = 120)
rolling_duplex_rate.df <- rolling_duplex_rate(duplex.df, simplex.df, type = "read")
}
\author{
Auden Block

Calculate the rolling average duplex nucleotide or read rate over an experiment at given intervals.
}
