% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_only.R
\name{simplex_only}
\alias{simplex_only}
\title{Find the simplex reads from which no duplex read was generated.}
\usage{
simplex_only(duplex, simplex)
}
\arguments{
\item{duplex.df}{A dataframe of duplex reads}

\item{simplex.df}{A dataframe of simplex reads of which the duplex reads spawned}
}
\value{
A dataframe of the simplex reads from which NO duplex reads were generated.
}
\description{
Find the simplex reads from which no duplex read was generated.
}
\examples{
simplex.only <- simplex.only(duplex.df, simplex.df);
summary.list$simplex.only <- simplex.only(summary.list$duplex.df, summary.list$simplex.df);
}
\author{
Auden Block
}
