% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_duplex_rate.R
\name{window_duplex_rate}
\alias{window_duplex_rate}
\title{Find the duplex rate at a specific timepoint.}
\usage{
window_duplex_rate(
  duplex,
  simplex,
  time_point,
  interval = 60,
  type = "nucleotide"
)
}
\arguments{
\item{time_point}{The timepoint at which to analyze in seconds}

\item{interval}{(Default = 60) The window before and after the timepoint to analyze in minutes.}

\item{type}{(Default = "nucleotide") The type of rate to be reported: nucleotide or read.}

\item{duplex.df}{A dataframe of duplex reads}

\item{simplex.df}{A dataframe of simplex reads}
}
\value{
A dataframe of (,1:3) the interval evaluated reported in seconds, minutes, and hours, and (,4) the duplex rate as a decimal during that time interval.
}
\description{
Find the duplex rate at a specific timepoint.
}
\examples{
window_duplex_rate(duplex.df, simplex.df, 3600, 60, type = "read")
}
\author{
Auden Block

#' Calculate the average duplex nucleotide or read rate at a specific timepoint and window size.
}
