% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_readIDs.R
\name{write_readIDs}
\alias{write_readIDs}
\title{Subset __X coverage using Nanopore Reads}
\usage{
write_readIDs(reads, file_name = "NULL")
}
\arguments{
\item{reads}{A dataframe containing the read_ids you wish to export.}

\item{file_name}{(Default = paste0(as.character(substitute(reads)),"_readIDs.txt")) Name for the file to be created.}
}
\value{
A txt file of readIDs.
}
\description{
Subset __X coverage using Nanopore Reads
}
\examples{
2x.subset <- nanopore_subset(reads, 2, 1500, TRUE)
}
\author{
Auden Block

Export the readIDs for a given dataframe of reads. 
This script is typically ran after nanopore_subset to save the randomly generated subset of readIDs for utilization in Shell/Bash.
}
