% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\title{Subset __X coverage using Nanopore Reads}
\usage{
coverage(reads, genome_size, units = "gb")
}
\arguments{
\item{reads}{A vector or list of dataframes (for users with multiple libraries wishing to calculate the total coverage, the sequence lengths must be labelled "sequence_length_template").}

\item{genome_size}{Estimated genome size.}

\item{units}{is the genome in bases, kilobases, megabases, or gigabases (default = "gb")}
}
\value{
A vector of subset read_ids/indexes whose length is equal to the desired coverage of a given genome.
}
\description{
Subset __X coverage using Nanopore Reads
}
\examples{
2x.subset <- nanopore_subset(reads, 2, 1500, TRUE, 10)
}
\author{
Auden Block
}
