% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanopore_filter.R
\name{nanopore_filter}
\alias{nanopore_filter}
\title{Subset __X coverage using Nanopore Reads}
\usage{
nanopore_filter(reads, min_quality = 1, min_length = 200, remove = FALSE)
}
\arguments{
\item{reads}{A dataframe containing the column names: read_id and sequence_length_template.}

\item{coverage}{The desired coverage.}
}
\value{
A subset of the input duplex dataframe in which a read is used as both a template and a complement. Useful for finding false positive duplex reads
}
\description{
Subset __X coverage using Nanopore Reads
}
\examples{
library2_duplex_table <- dorado_table(simplex.df, duplex.df)
all.libraries.duplex.table <- as.data.frame(lapply(simplex.list, dorado_table))
}
\author{
Auden Block
}
