% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitysqcore_vs_readlength.R
\name{qscore_vs_readlength_graph}
\alias{qscore_vs_readlength_graph}
\title{Dataframe containing a summary table about a Dorado run.}
\usage{
qscore_vs_readlength_graph(
  simplex,
  histogram = TRUE,
  colors = c("#e76f51", "#2a9d8f"),
  qscore = 10,
  theme = theme_classic()
)
}
\arguments{
\item{simplex}{A simplex read dataframe.}

\item{duplex}{A duplex read dataframe from the same summary as the simplex dataframe.}
}
\value{
A subset of the input duplex dataframe in which a read is used as both a template and a complement. Useful for finding false positive duplex reads
}
\description{
Dataframe containing a summary table about a Dorado run.
}
\examples{
library2_duplex_table <- dorado_table(simplex.df, duplex.df)
all.libraries.duplex.table <- as.data.frame(lapply(simplex.list, dorado_table))
}
\author{
Auden Block
}
