% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_only.R
\name{simplex_only}
\alias{simplex_only}
\title{Simplex_only}
\usage{
simplex_only(duplex, simplex)
}
\arguments{
\item{duplex}{A vector or dataframe (containing the column "read_id" or c("complement_id", "template_id")) of duplex reads.}

\item{simplex}{A vector or dataframe of simplex reads including from those from which the duplex reads spawned.}
}
\value{
A character vector or subsetted dataframe of simplex reads from which NO duplex reads were generated.
}
\description{
Subset the simplex reads dataframe to contain only reads from which no duplex read was generated.
}
\examples{
simplex_only(summary.list$duplex.df, summary.list$simplex.df)
}
\author{
Auden Block
}
