% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fastq.R
\name{write_fastq}
\alias{write_fastq}
\title{Export the fastq file into R}
\usage{
write_fastq(fastq.df, path)
}
\arguments{
\item{fastq.df}{The f}
}
\value{
A fastq file of the fastq dataframe.
}
\description{
Export the fastq file into R
}
\examples{
write_fastq(fastq.df,"<path_to_file>")
}
\author{
Auden Block

This function is for those wishing to use R to manipulate and evaluate fastq files either through DuplexVisualized or other bioinformatics programs.
}
